Fixed imports after renaming 'restrictions' -> 'commons', imports restructurized
